chore(TCOMP-681): split Wizard to Create and Edit wizards
